requirements.txt
matplotlib
numpy
Pillow
scikit_image
